+++ package.json
@@ -106,9 +106,9 @@
     "@types/node": "^24.10.1",
     "@types/react": "^19.2.7",
     "@types/react-dom": "^19.2.3",
-    "@types/sharp": "^0.32.0",
     "@types/slug": "^5.0.9",
     "@vitest/ui": "^4.0.18",
+    "baseline-browser-mapping": "^2.9.19",
     "eslint": "^9.39.1",
     "eslint-config-next": "16.0.7",
     "tailwindcss": "^4.1.17",
