+++ src/app/api/admin/upload/route.ts
@@ -2,9 +2,9 @@ import { db } from "@/lib/db";
 import { songs } from "@/lib/db/schema-pg";
 import { logger } from "@/lib/logger";
 import { VersionId } from "@/lib/metadata";
-import { Difficulty, NoteCounts, Region, SongType } from "@/lib/types";
+import { Difficulty, Region, SongType } from "@/lib/types";
 import { UpdateSong } from "@/lib/types/update";
-import { mergeSongs, taker, merger, key } from "@/server/services/admin/level-fetcher";
+import { mergeSongs, taker, merger, key } from "@/server/services/admin/fetcher-utils";
 import { important, PendingSong, value, Pending } from "@/server/utils/admin/type";
 import { and, eq } from "drizzle-orm";
 import { NextRequest, NextResponse } from "next/server";
